Separate class in 2 files.
