+++ MvvX.CouchBaseLite/MvvX.Plugins.CouchBaseLite/Managers/JsonSerializationSettings.cs
@@ -1,36 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MvvX.Plugins.CouchBaseLite.Managers
 {
-    /// <summary>
-    /// Specifies methods for deserializing date times that are written as strings
-    /// </summary>
-    public enum DateTimeHandling
-    {
-        /// <summary>
-        /// Deserialize to System.DateTime (local time zone)
-        /// </summary>
-        UseDateTime,
-
-        /// <summary>
-        /// Deserialize to System.DateTimeOffset (embedded time zone)
-        /// </summary>
-        UseDateTimeOffset,
-
-        /// <summary>
-        /// Don't deserialize (keep as a date-time string).
-        /// </summary>
-        /// <remarks>
-        /// WARNING: This will cause any DateTime or DateTimeOffset values to be
-        /// returned as strings.  Make sure you know what you are doing
-        /// </remarks>
-        Ignore
-    }
-
     /// <summary>
     /// A struct containing options for JSON serialization and deserialization
     /// </summary>
